ahora OrderDto.totalAmount es BigDecimal

+++ ecommerce/src/main/java/com/stellaTech/ecommerce/service/dto/OrderDto.java
@@ -24,7 +24,7 @@ public class OrderDto<T extends OrderDto.OrderItemDto> {
 
     @Null(groups = ValidationGroup.OnInsert.class)
     @NotNull(groups = ValidationGroup.OnRead.class)
-    private Long totalAmount;
+    private BigDecimal totalAmount;
 
     @NotEmpty
     @Singular
